Fix issue of listing bootstrap options in list, show, delete.
