+++ lib/chef/knife/cloud/helpers.rb
@@ -32,6 +32,42 @@ class Chef
           config[key] || Chef::Config[:knife][key]
         end
 
+        def create_service_instance
+          raise Chef::Exceptions::Override, "You must override create_service_instance in #{self} to create cloud specific service"
+        end
+
+        def execute_command
+          raise Chef::Exceptions::Override, "You must override execute_command in #{self}"
+        end
+
+        # Derived classes can override before_exec_command and after_exec_command
+        def before_exec_command
+        end
+
+        def after_exec_command
+        end
+
+        def set_default_config
+        end
+
+        def validate!(*keys)
+          # validates necessary options/params to carry out the command.
+          # subclasses to implement this.
+          errors = []
+          keys.each do |k|
+            errors << "You did not provide a valid '#{pretty_key(k)}' value." if locate_config_value(k).nil?
+          end
+          error_message = ""
+          raise CloudExceptions::ValidationError, error_message if errors.each { |e| ui.error(e); error_message = "#{error_message} #{e}." }.any?
+        end
+
+        def validate_params!
+        end
+
+        def pretty_key(key)
+          key.to_s.tr("_", " ").gsub(/\w+/) { |w| (w =~ /(ssh)|(aws)/i) ? w.upcase : w.capitalize }
+        end
+
       end
     end
   end
